Move Python wrappers to runtime
