+++ src/runtime/Cargo.toml
@@ -25,6 +25,8 @@ linkml_meta = { path = "../metamodel" }
 rio_turtle = "0.8"
 rio_api = "0.8"
 pyo3 = { version = "0.25.0", optional = true }
+serde_yml = "0.0.12"
+serde_path_to_error = "0.1.17"
 
 [dev-dependencies]
 assert_cmd = "2"
@@ -32,5 +34,5 @@ predicates = "2"
 tempfile = "3"
 
 [features]
-python = ["pyo3", "linkml_schemaview/python", "linkml_meta/pyo3"]
+python = ["pyo3", "linkml_meta/pyo3"]
 extension-module = ["pyo3/extension-module"]
