+++ src/schemaview/Cargo.toml
@@ -6,15 +6,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
 name = "linkml_schemaview"
-crate-type = ["cdylib", "rlib"]
+crate-type = ["rlib"]
 
 [dependencies]
-pyo3 = { version = "0.25.0", optional = true }
 linkml_meta = { path = "../metamodel" }
 serde_yml = "0.0.12"
 serde_path_to_error = "0.1.17"
 reqwest = { version = "0.12.16" , features = ["blocking"] }
 curies = "0.1.3"
-
-[features]
-python = ["pyo3", "linkml_meta/pyo3"]
